requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-SQLAlchemy
gunicorn
openai
requests
moviepy
pytube
python-dotenv
yt_dlp
